LeaveTopFeature should be called before enterFeature + update map docs
